% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SaaSc.R
\name{computeCorrelation}
\alias{computeCorrelation}
\title{Compute correlation between signaling and response}
\usage{
computeCorrelation(
  object,
  response.data = NULL,
  signaling.data = NULL,
  signature = NULL,
  cytokine = NULL,
  threshold = 100
)
}
\arguments{
\item{object}{A Seurat object processed by rebuildMatrix, and include the sample in meta.data assay.}

\item{response.data}{Response data frame generated by computeResponse.}

\item{signaling.data}{Signaling data frame generated by computeSignaling.}

\item{signature}{Signature to be calculated.}

\item{cytokine}{Cytokine to be calculated.}

\item{threshold}{The cell number threshold for samples. Default set to 100.}
}
\value{
A data frame include 6 columns: sample, nCell, signature, cytokine, correlation, qvalue.
}
\description{
Compute correlation between signaling and response
}
