% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SaaSc.R
\name{computeResponse}
\alias{computeResponse}
\title{Compute response}
\usage{
computeResponse(
  object,
  gene.rate = NULL,
  celltype = NULL,
  signature = NULL,
  cell.threshold = 100
)
}
\arguments{
\item{object}{A Seurat object processed by rebuildMatrix.}

\item{gene.rate}{Gene rate matrix generated by getGeneRate.}

\item{celltype}{Celltype to be calculated.}

\item{signature}{Signature name corresponding to cell type.}

\item{cell.threshold}{The cell number threshold for this cell type. Default set to 100.}
}
\value{
A data frame store the signature response of single cells.
}
\description{
Compute response
}
