% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SaaSc.R
\name{computeSignaling}
\alias{computeSignaling}
\title{Compute signaling}
\usage{
computeSignaling(
  object,
  model.file = NULL,
  celltype = NULL,
  cytokine = NULL,
  lambda = 10000,
  num.permutations = 1000,
  test.method = "two-sided"
)
}
\arguments{
\item{object}{A Seurat object processed by rebuildMatrix.}

\item{model.file}{Cytokine effects on genes quantitative matrix file, which row is gene and column is cytokine.}

\item{celltype}{Celltypes to be calculated.}

\item{cytokine}{Cytokines to be calculated.}

\item{lambda}{Lambda values of ridge regression, default set to 10000.}

\item{num.permutations}{Number of permutations to use, if greater than 0, compute empirical t-values using a permutation test. Default set to 1000.}

\item{test.method}{Significance test method. The choices are "two-sided", "greater" and "less", default is "two-sided".}
}
\value{
A data frame store the cytokine signaling of single cells.
}
\description{
Compute signaling
}
