% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SaaSc.R
\name{doInteraction}
\alias{doInteraction}
\title{Do interaction of Tres model}
\usage{
doInteraction(
  object,
  response.data = NULL,
  signaling.data = NULL,
  signature = NULL,
  cytokine = NULL,
  threshold = 100
)
}
\arguments{
\item{object}{A Seurat object processed by rebuildMatrix.}

\item{response.data}{Response data frame generated by computeResponse.}

\item{signaling.data}{Signaling data frame generated by computeSignaling.}

\item{signature}{Signature to be calculated.}

\item{cytokine}{Cytokine to be calculated.}

\item{threshold}{The cell number threshold for samples. Default set to 100.}
}
\value{
A data frame include 6 columns: sample, nCell, signature, gene, t, qvalue.
}
\description{
Do interaction of Tres model
}
