% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SaaSc.R
\name{getGeneRate}
\alias{getGeneRate}
\title{Get gene occurrence rate in background and signature genesets}
\usage{
getGeneRate(
  background.geneset = NULL,
  signature.geneset = NULL,
  mode = "single"
)
}
\arguments{
\item{background.geneset}{Background genesets file name (in GMT format) or R list object.}

\item{signature.geneset}{Signature geneset, a R object of list which include tags (signature name) and values (gene list).}

\item{mode}{Signature geneset mode, "single" means each geneset is a signature, "multiple" means all genesets form a signature.}
}
\value{
A data frame include background and signatures gene rate.
}
\description{
Get gene occurrence rate in background and signature genesets
}
