% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SaaSc.R
\name{getTresSignature}
\alias{getTresSignature}
\title{Get Tres signature}
\usage{
getTresSignature(
  interaction.dataset = NULL,
  signature.cytokine = NULL,
  qvalue = 0.05,
  cutoff = 0.5,
  method = "median"
)
}
\arguments{
\item{interaction.dataset}{Interaction dataset of all training samples, which generated by computeResponse and aggregated.}

\item{signature.cytokine}{Cytokine to be calculated.}

\item{qvalue}{q value threshold, default is 0.05.}

\item{cutoff}{Sample proportion threshold less than q value, default set to 0.5.}

\item{method}{Method for summarizing Tres signatures when the number of signature.cytokine more than one. The choices include "medium" and "mean", default set to "medium."}
}
\value{
Tres signature data frame include Tres score of valid genes.
}
\description{
Get Tres signature
}
