% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SaaSc.R
\name{rebuildMatrix}
\alias{rebuildMatrix}
\title{Rebuild the gene expression matrix}
\usage{
rebuildMatrix(
  object,
  features = NULL,
  assay = "RNA",
  slot = "data",
  method = NULL,
  dim = NULL,
  avc = NULL
)
}
\arguments{
\item{object}{A seurat object, must include the celltype in meta.data assay.}

\item{features}{Character vector of feature names. If not specified all features will be taken.}

\item{assay}{Name of assay MCA is being run on, default set to "RNA".}

\item{slot}{The slot to pull expression data for Seurat, default set to "data".}

\item{method}{Dimensionality reduction and reconstruction method. The choices are NULL, "MCA" , "NMF" and "PCA". NULL means no reconstruction.}

\item{dim}{The number of selected dimensions, 1 <= dim <= 50. Enter only one of avc and dim.}

\item{avc}{Accumulative variance contribution threshold, 0 <= avc <= 1. Enter only one of avc and dim.}
}
\value{
A new seurat object include the 'rebuild.RNA' assay that store the reconstructed matrix.
}
\description{
Rebuild the gene expression matrix
}
