% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SaaSc.R
\name{computeResponse}
\alias{computeResponse}
\title{Compute response}
\usage{
computeResponse(object, gene.rate = NULL, celltype = NULL, signature = NULL)
}
\arguments{
\item{object}{A Seurat object processed by rebuildMatrix.}

\item{gene.rate}{Gene rate matrix generated by getGeneRate.}

\item{celltype}{A cell type to be calculated.}

\item{signature}{Signature name corresponding to the cell type.}
}
\value{
A data frame store the signature response of single cells.
}
\description{
Compute response
}
