% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SaaSc.R
\name{doInteraction}
\alias{doInteraction}
\title{Do interaction of Tres model}
\usage{
doInteraction(
  object,
  method = NULL,
  dim = NULL,
  avc = NULL,
  response.data = NULL,
  signaling.data = NULL,
  signature = NULL,
  cytokine = NULL,
  threshold = 100
)
}
\arguments{
\item{object}{A Seurat object processed by rebuildMatrix().}

\item{method}{Dimensionality reduction and reconstruction method, in a character. The choices are NULL(do not rebuild), "mca" , "nmf" and "pca".}

\item{dim}{The number of selected dimensions, 1 <= dim <= 50. Enter only one of avc and dim.}

\item{avc}{Accumulative variance contribution threshold, 0 <= avc <= 1. Enter only one of avc and dim.}

\item{response.data}{Response data frame generated by computeResponse().}

\item{signaling.data}{Signaling data frame generated by computeSignaling().}

\item{signature}{Signature to be calculated.}

\item{cytokine}{Cytokine to be calculated.}

\item{threshold}{The cell number threshold for samples. Default set to 100.}
}
\value{
A data frame include 6 columns: sample, nCell, signature, gene, t, qvalue.
}
\description{
Do interaction of Tres model
}
