% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SaaSc.R
\name{rebuildMatrix}
\alias{rebuildMatrix}
\title{Rebuild the gene expression matrix}
\usage{
rebuildMatrix(
  object,
  features = NULL,
  assay = "RNA",
  slot = "data",
  methods = NULL
)
}
\arguments{
\item{object}{A seurat object, must include the celltype in meta.data assay.}

\item{features}{Character vector of feature names. If not specified all features will be taken.}

\item{assay}{Name of assay MCA is being run on, default set to "RNA".}

\item{slot}{The slot to pull expression data for Seurat, default set to "data".}

\item{methods}{Dimensionality reduction and reconstruction methods, in a character list. The choices are "mca" , "nmf" and "pca".}
}
\value{
A new seurat object include the 'rebuild.RNA' assay that store the reconstructed matrix.
}
\description{
Rebuild the gene expression matrix
}
